+++ src/pytorch_lightning/core/mixins/__init__.py
@@ -12,5 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from pytorch_lightning.core.mixins.device_dtype_mixin import DeviceDtypeModuleMixin  # noqa: F401
+from lightning_lite.utilities.device_dtype_mixin import DeviceDtypeModuleMixin  # noqa: F401
 from pytorch_lightning.core.mixins.hparams_mixin import HyperparametersMixin  # noqa: F401
